refactor api functions file structure
