+++ src/utilities/api/config.js
@@ -1,0 +1,5 @@
+import axios from "axios";
+
+export const api = axios.create({
+  baseURL: "https://be-nc-news-8igb.onrender.com/api",
+});
