+++ src/utilities/api/index.js
@@ -1,0 +1,3 @@
+export * from "./articles";
+export * from "./comments";
+export * from "./users";
