+++ src/utilities/api/users.js
@@ -1,0 +1,5 @@
+import { api } from "./config";
+
+export function getUser(username) {
+  return api.get(`/users/${username}`).then(({ data }) => data.user);
+}
